Fix `WaitRecoveryDone` to actually wait for server to exit recovery (#5)
Waiting for the `recovery.signal` file to disappear is NOT the correct
way to determine if the server exited recovery or not, because the file
is removed shortly _before_ Postgres formally exits recovery mode.  So
with timings just right, Stolon would stop the server before recovery
was recorded as done, resulting in

    database system was shut down in recovery
    database system was not properly shut down; automatic recovery in progress

upon restarting the server.  In a trivial recovery (to the latest
consistent state) that wouldn't matter, but any and all of the
`recovery_target_` settings would effectively be ignored as Stolon then
restarts Postgres in normal non-recovery mode which replays WAL to the
end.

Fix this by doing what `pg_ctl -w` does: look at `postmaster.pid`'s
status line and wait for it to become `running`.